test(ormus): move handler tests to separate files
